Move functions and feedback.
