+++ music_assistant/providers/_demo_audio_analysis_provider/__init__.py
@@ -96,7 +96,7 @@ class DemoAudioAnalysisProvider(AudioAnalysisProvider):
       The base class uses this to skip re-analysis of already-analyzed tracks.
     - If you have other conditions that determine whether to skip an analysis,
       implement them in _start_analysis and return False to reject the session.
-    - Store results via self.mass.music.set_audio_analysis() in _finalize.
+    - Store results via self.mass.streams.audio_analysis.set_audio_analysis() in _finalize.
     """
 
     # Increment this when your analysis algorithm changes significantly.
@@ -164,9 +164,9 @@ class DemoAudioAnalysisProvider(AudioAnalysisProvider):
     async def _finalize(self, session_id: str) -> None:
         """Finalize analysis and store results.
 
-        Called when the track has finished streaming and all chunks have been
+        Called when the track has finished buffering and all chunks have been
         processed. This is where a real provider would compute final results
-        and store them via self.mass.music.set_audio_analysis().
+        and store them via self.mass.streams.audio_analysis.set_audio_analysis().
 
         Example of storing results (not done in this demo)::
 
@@ -174,7 +174,7 @@ class DemoAudioAnalysisProvider(AudioAnalysisProvider):
 
             session = self._sessions[session_id]
             analysis = AudioAnalysisData(bpm=120.0, duration=180.5)
-            await self.mass.music.set_audio_analysis(
+            await self.mass.streams.audio_analysis.set_audio_analysis(
                 item_id=session.streamdetails.item_id,
                 provider_instance_id_or_domain=session.streamdetails.provider,
                 aa_provider_domain=self.domain,
